(PC-18695) feat(bookings): move all pages in specific folders
