+++ src/features/navigation/RootNavigator/routes.ts
@@ -16,8 +16,8 @@ import { AccountReactivationSuccess } from 'features/auth/suspendedAccount/Accou
 import { FraudulentAccount } from 'features/auth/suspendedAccount/FraudulentAccount/FraudulentAccount'
 import { SuspendedAccount } from 'features/auth/suspendedAccount/SuspendedAccount/SuspendedAccount'
 import { SuspensionScreen } from 'features/auth/suspendedAccount/SuspensionScreen/SuspensionScreen'
-import { BookingDetails } from 'features/bookings/pages/BookingDetails'
-import { EndedBookings } from 'features/bookings/pages/EndedBookings'
+import { BookingDetails } from 'features/bookings/pages/BookingDetails/BookingDetails'
+import { EndedBookings } from 'features/bookings/pages/EndedBookings/EndedBookings'
 import { BookingConfirmation } from 'features/bookOffer/pages/BookingConfirmation'
 import { AppComponents } from 'features/cheatcodes/pages/AppComponents/AppComponents'
 import { CheatCodes } from 'features/cheatcodes/pages/CheatCodes/CheatCodes'
